Massively improve inline toolchain build
build.sh is now sourced from envsetup.sh,
so these functions may be used independently of
one another, or the full build function may be run
on it's own for arm-linux-androideabi currently.

If the build does not complete, the user gets
a nice failure message, and requests they check
scrollback.

This makes support for new toolchain triplets far easier,
as well as toolchain testing.